pyproject.toml
[project]
name = "provisioning-station"
version = "1.0.0"
description = "IoT Solution Provisioning Platform for Seeed Studio products"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Seeed Studio", email = "support@seeedstudio.com"}
]
keywords = ["iot", "provisioning", "seeed", "esp32", "docker", "deployment"]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pyyaml>=6.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "esptool>=4.7.0",
    "pyserial>=3.5",
    "paramiko>=3.4.0",
    "docker>=7.0.0",
    "websockets>=12.0",
    "aiofiles>=23.2.1",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "markdown>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
]

[project.scripts]
provisioning-station = "provisioning_station.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["provisioning_station"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
